fix: change incorrect login URL
Django re-writes all requests that do not have a slash at the end to
include a slash, but for some reason the request type is discarded, and
we see the magic of a POST request becoming a GET request. Of course,
Django REST Framework is none too happy about the request and gives back
a bogus result. (TODO: handle that in the next commit)