fix: use urllib for joining URLs
The previous implementation just blindly combined the two, which could
result in the following:

  https://server.example.com//api/v1/...

if the base server URL supplied by the user had a trailing slash at the
end.

The dirty solution (at the time) was to disallow trailign slashes in the
base server URL, but with this change we should be able to allow slashes
in the base server URL again.